Move move effect code to its own file
