+++ packages/dds/tree/src/feature-libraries/sequence-field/index.ts
@@ -62,4 +62,10 @@ export { MarkListFactory } from "./markListFactory";
 export { NodeChangeRebaser, rebase } from "./rebase";
 export { invert, NodeChangeInverter } from "./invert";
 export { compose, NodeChangeComposer } from "./compose";
-export { isMoveMark, MoveMark, MoveEffectTable, MovePartition, newMoveEffectTable } from "./utils";
+export {
+    isMoveMark,
+    MoveMark,
+    MoveEffectTable,
+    MovePartition,
+    newMoveEffectTable,
+} from "./moveEffectTable";
